file-utilities: fix name length clamping call type
The compiler converted the signed integer to unsigned, as strlen()
returns an unsigned integer, which means the assertion of a minimum
length of zero could have gotten optimized out.

This fixes un<>signed comparison warnings.

+++ src/nautilus-file-utilities.c
@@ -1031,7 +1031,7 @@ nautilus_get_max_child_name_length_for_location (GFile *location)
         max_child_name_length = name_max;
         if (path_max != -1)
         {
-            max_child_name_length = CLAMP ((path_max - 1) - strlen (path),
+            max_child_name_length = CLAMP ((path_max - 1) - (glong) strlen (path),
                                            0,
                                            max_child_name_length);
         }
